Separe Group Shape Handling and Integrate Functionality
